Unrooted NNI comparisons (#24)

+++ src/TreeTools_splits.cpp
@@ -14,7 +14,7 @@ RawMatrix cpp_edge_to_splits(IntegerMatrix edge, IntegerVector nTip) {
   if (edge.cols() != 2) {
     throw std::invalid_argument("Edge matrix must contain two columns");
   }
-  if (1L + edge.rows() > UINTX_MAX - 1) { /* UINT_MAX denotes NOT_TRIVAL */
+  if (1L + edge.rows() > long(UINTX_MAX - 1)) { /* UINT_MAX denotes NOT_TRIVAL */
     throw(std::length_error("Too many edges in tree for edge_to_splits: "       // # nocov
                               "Contact maintainer for advice"));                // # nocov
   }
